% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(data_file = "data.csv", user = "", password = "", form_id = "saint")
}
\arguments{
\item{data_file}{The data file}

\item{user}{The user}

\item{password}{The password}

\item{form_id}{The form id}
}
\value{
Data is read
}
\description{
Retrieve data
}
