% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_aws.R
\name{get_data_aws}
\alias{get_data_aws}
\title{Get data AWS}
\usage{
get_data_aws(s3_csv_path)
}
\arguments{
\item{s3_csv_path}{Path to an S3 creds csv}
}
\value{
Data is read
}
\description{
Retrieve data from AWS
}
