% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_aws.R
\name{update_data_aws}
\alias{update_data_aws}
\title{Update data AWS}
\usage{
update_data_aws(s3_csv_path, df, creds)
}
\arguments{
\item{s3_csv_path}{Path to an S3 creds csv}

\item{df}{A data frame}

\item{creds}{A credentials yaml path}
}
\value{
Data is read
}
\description{
Write data to AWS
}
