% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_aws.R
\name{write_data_aws}
\alias{write_data_aws}
\title{Write data AWS}
\usage{
write_data_aws(s3_csv_path, df)
}
\arguments{
\item{s3_csv_path}{Path to an S3 creds csv}

\item{df}{A data frame}
}
\value{
Data is read
}
\description{
Write data to AWS
}
